refactor: improve VM and smart contract modules
Enhance VM execution engine and smart contract functionality with better
error handling and compatibility improvements.

VM improvements:
- Enhanced execution engine with better error propagation
- Improved evaluation stack operations
- Better interop service handling
- Optimized jump table operations
- Enhanced script validation

Smart Contract improvements:
- Better application engine integration
- Improved native contract implementations
- Enhanced contract manifest handling
- Better event emission system
- Governance types addition

+++ crates/vm/src/application_engine.rs
@@ -937,6 +937,7 @@ impl From<ApplicationEngine> for ExecutionEngine {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use num_bigint::BigInt;
 
     #[test]
     fn test_application_engine_creation() {
@@ -1029,7 +1030,7 @@ mod tests {
                     .expect("operation should succeed")
                     .as_int()
                     .expect("Operation failed"),
-                3
+                BigInt::from(3)
             );
         } else {
             panic!("Execution failed: {state:?}");
